% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_weighted_sbm.R
\name{gen_weighted_sbm}
\alias{gen_weighted_sbm}
\title{Generate a Weighted Stochastic Block Model (SBM) Network}
\usage{
gen_weighted_sbm(
  node_num = 200,
  cluster_size = c(50, 50, 50, 50),
  win_cluster_den = 0.3,
  win_cluster_dist = "runif",
  win_cluster_par = list(min = 1, max = 2),
  btw_cluster_den = 0.1,
  btw_cluster_dist = "runif",
  btw_cluster_par = list(min = 1, max = 2)
)
}
\arguments{
\item{node_num}{Total number of nodes. Default is 200.}

\item{cluster_size}{A vector of community sizes. Must sum to \code{node_num}.}

\item{win_cluster_den}{Probability of edge existence within a community. Default is 0.3.}

\item{win_cluster_dist}{Distribution function for within-community edge weights (as string). Default is \code{"runif"}.}

\item{win_cluster_par}{List of parameters for \code{win_cluster_dist}. Default is \code{list(min = 1, max = 2)}.}

\item{btw_cluster_den}{Probability of edge existence between communities. Default is 0.1.}

\item{btw_cluster_dist}{Distribution function for between-community edge weights (as string). Default is \code{"runif"}.}

\item{btw_cluster_par}{List of parameters for \code{btw_cluster_dist}. Default is \code{list(min = 1, max = 2)}.}
}
\value{
A list containing:
\item{node_number}{Total number of nodes.}
\item{cluster_number}{Number of clusters.}
\item{adjacency_matrix}{Symmetric weighted adjacency matrix.}
\item{network_sparsity}{Proportion of absent edges in the network.}
}
\description{
Simulate an undirected weighted network from a stochastic block model (SBM).
Within-community and between-community edges are drawn with specified
probabilities and weight distributions.
}
\examples{
set.seed(123)
net <- gen_weighted_sbm(
  node_num = 20,
  cluster_size = c(10, 10),
  win_cluster_den = 0.5,
  btw_cluster_den = 0.1
)
 W <- net$adjacency_matrix
 W

}
