% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASC.R
\name{CASC}
\alias{CASC}
\title{Covariate-Assisted Spectral Clustering (CASC)}
\usage{
CASC(Adj, Covariate, K, n = 5, ...)
}
\arguments{
\item{Adj}{An adjacency matrix of the network (symmetric, non-negative, with zero diagonal).}

\item{Covariate}{A node covariate matrix where rows correspond to nodes
and columns to covariates.}

\item{K}{The number of clusters to partition the nodes into (integer).}

\item{n}{Number of \eqn{\alpha} values to evaluate in the grid search range. Default is 5.}

\item{...}{Additional arguments passed to \code{kmeans()}, such as
\code{nstart}, \code{iter.max}, or initialization settings.}
}
\value{
An object of class \code{"CASC"} containing:
\item{alpha}{The selected tuning parameter value.}
\item{K}{The number of clusters.}
\item{cluster}{Cluster membership assignments for each node.}
\item{alphalower}{Lower bound of the alpha search grid.}
\item{alphaupper}{Upper bound of the alpha search grid.}
}
\description{
Perform covariate-assisted spectral clustering (CASC) that integrates
network connectivity and node covariates. The method searches over a
range of tuning parameters (\eqn{\alpha}) to balance structural and
covariate information and selects the best clustering result.
}
\details{
Covariate-assisted spectral clustering (CASC) is a community detection
algorithm for networks with node covariates, proposed by
Binkiewicz, Vogelstein, and Rohe (2017).
Let \eqn{X \in \mathbb{R}^{n \times p}} be the covariate matrix,
where each row corresponds to a node’s covariate vector.
Define the regularized graph Laplacian as
\deqn{L_\tau = D_\tau^{-1/2} A D_\tau^{-1/2},}
where \eqn{D_\tau = D + \tau I} and \eqn{D} is the diagonal degree matrix of Adj matrix \eqn{A}.
CASC then forms a balanced matrix:
\deqn{\tilde{L}(\alpha) = L_\tau L_\tau + \alpha X X^\top,}
where \eqn{\alpha} is a tuning parameter controlling the relative weight
of covariate information.
Clustering is performed by applying k-means on the first \eqn{K}
leading eigenvectors of \eqn{\tilde{L}(\alpha)}.
}
\examples{
library(MASS)
set.seed(123)

n <- 200; k <- 4; d <- 10; se <- 2
z <- rep(1:k, each = n/k)

# --- SBM adjacency ---
p <- matrix(0.2, k, k); diag(p) <- 0.3
A <- (matrix(runif(n^2), n, n) < p[z, z]) * 1
A[lower.tri(A)] <- t(A)[lower.tri(A)]
diag(A) <- 0

# --- Covariates ---
means <- matrix(c(
  rep(d / sqrt(8), 3),
  rep(c(-d, -d, d) / sqrt(8), 1),
  rep(c(-d, d, -d) / sqrt(8), 1),
  rep(c(d, -d, -d) / sqrt(8), 1)
), nrow = k, byrow = TRUE)
X <- do.call(rbind, lapply(1:k, \(i) MASS::mvrnorm(n/k, means[i, ], diag(se^2, 3))))

# Run CASC clustering
result_CASC <- CASC(A, X, K = 4, iter.max = 100, nstart = 10)
result_CASC

}
\references{
Binkiewicz, N., Vogelstein, J. T., & Rohe, K. (2017).
Covariate-assisted spectral clustering. \emph{Biometrika}, 104(2), 361–377.
}
